VCrypture-API/requirements.txt
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Executor==0.9.3
Flask-Mail==0.9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pycryptodome==3.9.8
requests==2.24.0
urllib3==1.25.10
Werkzeug==1.0.1
